// src/lib/authoring/types.ts
export interface CategoryData {
  name: string;
  slug: string;
  description: string;
  icon: string;
  color: string;
}

export interface SubcategoryData {
  name: string;
  slug: string;
  category: string; // parent category slug
  description: string;
}

export interface TagData {
  name: string;
  slug: string;
  description: string;
  status: 'active' | 'deprecated' | 'merged';
  mergedIntoSlug?: string;
}

export interface AuthorData {
  name: string;
  slug: string;
  gender?: 'male' | 'female';
  role?: string;
  bio?: string;
  avatar?: string;
}